requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==6.0.0
anomalib==2.2.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
docstring_parser==0.17.0
einops==0.8.1
filelock==3.20.0
fonttools==4.61.0
FrEIA==0.2
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.1
idna==3.11
imagecodecs==2025.11.11
ImageIO==2.37.2
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.2
jsonargparse==4.44.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kornia==0.8.2
kornia_rs==0.1.10
lazy_loader==0.4
lightning==2.6.0
lightning-utilities==0.15.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
narwhals==2.13.0
networkx==3.6
numpy==2.2.6
omegaconf==2.3.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
protobuf==6.33.2
pyarrow==22.0.0
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytorch-lightning==2.6.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rich-argparse==1.7.2
rpds-py==0.30.0
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
streamlit==1.52.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.10.16
timm==1.0.22
toml==0.10.2
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchmetrics==1.8.2
torchvision==0.22.1+cu118
tornado==6.5.2
tqdm==4.67.1
typer-slim==0.20.0
typeshed_client==2.8.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
watchdog==6.0.0
yarl==1.22.0